Fix FreeBSD suport
Ref https://github.com/containers/skopeo/issues/1163

+++ pkg/archive/archive_freebsd.go
@@ -1,4 +1,4 @@
-// +build !windows,!freebsd
+// +build freebsd
 
 package archive
 
@@ -103,7 +103,7 @@ func handleTarTypeBlockCharFifo(hdr *tar.Header, path string) error {
 		mode |= unix.S_IFIFO
 	}
 
-	return system.Mknod(path, mode, int(system.Mkdev(hdr.Devmajor, hdr.Devminor)))
+	return system.Mknod(path, mode, uint64(system.Mkdev(hdr.Devmajor, hdr.Devminor)))
 }
 
 func handleLChmod(hdr *tar.Header, path string, hdrInfo os.FileInfo, forceMask *os.FileMode) error {
